+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## Changelog
 
-### 0.1.3.1
-* updated assets
+### Unreleased
+* added "From" before price
+* added input-text class to domain name field
+* fix options not included in cart total
 
-### 0.1.3
+### 0.1.3.1
 * renamed as woocommerce-domain-names (hopefully the last rename, to match my other wc plugins naming)
+* updated assets
 
 ### 0.1.2
 * added package updater